requirements.txt
PyMySQL==1.0.2
Flask-Session==0.4
black==21.9b0
SQLAlchemy==1.4.25
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.15.1
email_validator===1.1.3